правит открытие попапа
